triage_env/server/requirements.txt
fastapi
uvicorn
openenv-core>=0.1.0
pydantic>=2.0.0
